+++ build_all_mkdocs.sh
@@ -300,6 +300,14 @@ cd $root_dir/MkDocs/$lab
 mkdocs build
 echo "BUILD_INFO - The $lab lab is build and added under the top level of IBM Maximo Labs."
 
+# Build the Monitor Role Based Access Control 9.1 level:
+lab=monitor_role_based_access_control_9.1
+echo "==================================================================================="
+echo "BUILD_INFO - The $lab lab is being build"
+cd $root_dir/MkDocs/$lab
+mkdocs build
+echo "BUILD_INFO - The $lab lab is build and added under the top level of IBM Maximo Labs."
+
 # Build the Monitor Managed Gatewway Modbus 9.0 level:
 lab=monitor_managed_hierarchy_9.1
 echo "==================================================================================="